Partial fix of #17

+++ src/mannkendall/mk_tools.py
@@ -96,7 +96,8 @@ def nb_tie(data, resolution):
     # Avoid the use of np.arange because of floating point errors, in favor of linspace
     #bins = np.arange(np.nanmin(data), np.nanmax(data)+resolution, resolution)
     nbins = int((np.nanmax(data)-np.nanmin(data))//resolution + 1)
-    bins = np.linspace(np.nanmin(data), np.nanmin(data) + nbins * resolution, num=nbins + 1)
+    bins = np.linspace(np.nanmin(data), np.nanmin(data) + nbins * resolution, num=nbins + 1,
+                       dtype='float128') # Partial fix of #17
 
     # A sanity check
     if len(bins) < 2:
